Add MVT positioning logic
This adds functionality for a MVT test.
It will check the document for a MVT positioning div and move the
component inside it. This functionality was previously done inside
next-article.
The component now starts in a hidden state, is moved if required
and then made visible. Tweaked visibility toggle func to fit this.